backend/pyproject.toml
[project]
name = "gemini-tales"
version = "0.1.0"
description = "Backend for AI storytelling with multi-agent system"
authors = [
    {name = "Veronika Kashtanova", homepage = "https://x.com/veron_code"},
]
dependencies = [
    "google-adk==1.22.0",
    "opentelemetry-exporter-gcp-trace==1.11.0",
    "google-cloud-logging==3.13.0",
    "google-cloud-aiplatform[evaluation]==1.132.0",
    "fastapi==0.123.*",
    "uvicorn==0.40.0",
    "psycopg2-binary==2.9.*",
    "a2a-sdk==0.3.*",
    "websockets>=12.0",
    "google-genai>=1.64.0",
    "google-auth>=2.23.0",
    "certifi>=2023.7.22",
    "aiohttp>=3.8.0",
    "Pillow>=10.0.0",
    "httpx>=0.25.0",
    "httpx-sse>=0.4.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

requires-python = ">=3.10,<3.14"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "shared"]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true
